+++ cmn-data-svc-server/src/main/java/org/acumos/cds/domain/MLPSolutionFOM.java
@@ -57,7 +57,7 @@ public class MLPSolutionFOM extends MLPAbstractSolution implements Serializable 
 	 * serialized as JSON).
 	 */
 	@OneToOne
-	@JoinColumn(name = "OWNER_ID", nullable = false, columnDefinition = "CHAR(36)")
+	@JoinColumn(name = "USER_ID", nullable = false, columnDefinition = "CHAR(36)")
 	private MLPUser owner;
 
 	/**