+++ cmn-data-svc-server/src/main/java/org/acumos/cds/domain/MLPSolutionRevisionFOM.java
@@ -67,7 +67,7 @@ public class MLPSolutionRevisionFOM extends MLPAbstractSolutionRevision implemen
 	 * serialized as JSON).
 	 */
 	@OneToOne
-	@JoinColumn(name = "OWNER_ID", nullable = false, columnDefinition = "CHAR(36)")
+	@JoinColumn(name = "USER_ID", nullable = false, columnDefinition = "CHAR(36)")
 	private MLPUser owner;
 
 	/**